+++ src/lib/classes/automaton.ts
@@ -16,22 +16,25 @@ export type {
 
 // Enums
 export { Backend } from './automaton/backend';
+export { DeclarationType } from './automaton/declaration_type';
 export { Status } from './automaton/status';
 export { LocationType } from './automaton/location_type';
 export { OperatorType } from './automaton/operator_type';
+export { PropertyType } from './automaton/property_type';
 export { Urgency } from './automaton/urgency';
 
 // Minor Classes
 export { Invariant } from './automaton/invariant';
 export { Nickname } from './automaton/nickname';
 export { Operator } from './automaton/operator';
+export { Property } from './automaton/property';
 export { SystemEdge } from './automaton/system_edge';
 export { ComponentInstance } from './automaton/component_instance';
 
 // Major Classes
 export { Declaration } from './automaton/declaration';
 export { Component } from './automaton/component';
-export { Edge, Property, PropertyType } from './automaton/edge';
+export { Edge } from './automaton/edge';
 export { Location } from './automaton/location';
 export { Query, Queries } from './automaton/query';
 export { System } from './automaton/system';
