+++ src/lib/classes/automaton/backend.ts
@@ -1,4 +1,9 @@
 
+/**
+* What backend is targeted 
+*  - j-ECDAR or 
+*  - Reveaal
+* */
 export enum Backend {
 	J_ECDAR = 0,
 	REVEAAL = 1
