+++ src/lib/classes/automaton/component_instance.ts
@@ -2,9 +2,25 @@ import type {
   Point 
 } from '$lib/classes/draw';
 
+/**
+* # A Component Instance 
+* Is used by a System to store what components are used
+* */
 export class ComponentInstance {
+	/**
+	* The id of the component instance 
+	* This must be unique and is shared by Operators
+	* */
 	id: number;
+
+	/**
+	* The name of the component this references
+	* */
 	name: string;
+
+	/**
+	* The position of the component instance
+	* */
 	position: Point;
 
 	constructor(id: number, name: string, position: Point) {
