+++ src/lib/classes/automaton/declaration.ts
@@ -1,22 +1,26 @@
-import type { 
-  SerializeRaw,
-  ToRaw,
-  FromRaw,
-  DeserializeRaw,
-  RawDeclaration
+import { 
+  type SerializeRaw,
+  type ToRaw,
+  type FromRaw,
+  type DeserializeRaw,
+  type RawDeclaration,
+  DeclarationType
 } from '../automaton'
 
+/**
+* The top level declarations
+* */
 export class Declaration implements SerializeRaw, ToRaw<RawDeclaration> {
-	name: string;
+	type: DeclarationType;
 	declarations: string;
 
-	constructor(name = '', declarations = '') {
-		this.name = name;
+	constructor(type = DeclarationType.GLOBAL, declarations = "") {
+		this.type = type;
 		this.declarations = declarations;
 	}
 
 	static fromRaw: FromRaw<RawDeclaration, Declaration> = (raw) => {
-		return new Declaration(raw.name, raw.declarations);
+		return new Declaration(raw.name as DeclarationType, raw.declarations);
 	};
 
 	static deserializeRaw: DeserializeRaw<Declaration> = (input) => {
@@ -26,7 +30,7 @@ export class Declaration implements SerializeRaw, ToRaw<RawDeclaration> {
 
 	toRaw() {
 		return {
-			name: this.name,
+			name: this.type,
 			declarations: this.declarations
 		};
 	};
