+++ src/lib/classes/automaton/declaration_type.ts
@@ -1,0 +1,4 @@
+export enum DeclarationType {
+  GLOBAL = "Global Declarations",
+  SYSETEM = "System Declarations"
+}
