+++ src/lib/classes/automaton/invariant.ts
@@ -4,7 +4,15 @@ import type {
 } from '$lib/classes/draw';
 
 export class Invariant {
+	/**
+	* The invariant function
+	* ex c >= 8
+	* */
 	fn: string;
+
+	/**
+	* The position of the invariant
+	* */
 	position: Point;
 	constructor(fn: string, position: Point) {
 		this.fn = fn;
